added GET/ servers
At this stage I am not sure what the best practices for Flask are. Neither am I sure on how to structure this project.
There may be many revisions to the structure, but I will have time to figure that all out.
Progress on this project is bottlenecked by the frontend, so I will only be expanding this service as endpoints are needed by the frontend.