extracted ServerAudio from AudioClient
I am not sure whether this is a pattern or not, but it seems like modelling each resource I intend to serve RESTfully might be the move.
Everything I am doing is basically wrapping Discord.py to expose an API my interface can talk to, so I don't imagine I need to do anything
complicated in this project - just a lot of plumbing really.