interface request: serialize id as str
Current ids are larger than JS can parse, which leads to the interface
losing a number of digits from the id when mapping it to a property. Fatal issue. The only solutions would be to use a library
to parse the JSON on the frontend in a special way, or to amend the response serialization.